+++ openvpn/ssl/protostack.hpp
@@ -313,7 +313,7 @@ namespace openvpn {
 		  error(Error::SSL_ERROR);
 		  throw;
 		}
-	      if (size == buf->size())
+	      if (size == static_cast<ssize_t>(buf->size()))
 		app_write_queue.pop_front();
 	      else if (size == SSLConst::SHOULD_RETRY)
 		break;
